+++ pycurd/crud/base_crud.py
@@ -6,7 +6,7 @@ import pydantic
 
 from pycurd.const import QUERY_OP_RELATION
 from pycurd.crud._core_crud import CoreCrud
-from pycurd.crud.query_result_row import QueryResultRow
+from pycurd.crud.query_result_row import QueryResultRow, QueryResultRowList
 from pycurd.error import PermissionException
 from pycurd.permission import RoleDefine, A
 from pycurd.query import QueryInfo, QueryConditions, ConditionExpr, QueryJoinInfo, ConditionLogicExpr
@@ -122,7 +122,6 @@ class BaseCrud(CoreCrud, ABC):
             # TODO: 权限检查之后过不了检验的后面再处置
             pass
 
-
         if not values:
             return []
 
@@ -142,15 +141,15 @@ class BaseCrud(CoreCrud, ABC):
         info = await self._solve_query(info, perm)
         return await self.delete(info, _perm=perm)
 
-    async def get_list_with_perm(self, info: QueryInfo, with_count=False, *, perm: PermInfo) -> List[QueryResultRow]:
+    async def get_list_with_perm(self, info: QueryInfo, with_count=False, *, perm: PermInfo) -> QueryResultRowList:
         info = await self._solve_query(info, perm)
         return await self.get_list(info, with_count, _perm=perm)
 
-    async def get_list_with_foreign_keys(self, info: QueryInfo, perm: PermInfo = None):
+    async def get_list_with_foreign_keys(self, info: QueryInfo, with_count=False, perm: PermInfo = None) -> QueryResultRowList:
         if perm is None:
             perm = PermInfo(False, None, None)
 
-        ret = await self.get_list_with_perm(info, perm=perm)
+        ret = await self.get_list_with_perm(info, with_count, perm=perm)
 
         async def solve(ret_lst, main_table, fk_queries, depth=0):
             if fk_queries is None: