+++ changelog.md
@@ -1,3 +1,22 @@
+### 0.1.7 update 2020.11.03
 
-#### 0.1.15 update 2020.11.02
+* Fixed: operator: contains, in, notin
 
+* Fixed: no insert return for psycopg2
+
+* Fixed: incorrect placeholder for psycopg2 and peewee
+
+
+#### 0.1.6 update 2020.11.02
+
+* Added: tortoise supported
+
+* Added: PostgreSQL array field supported
+
+* Added: json field supported
+
+* Fixed: insert blob failed
+
+* Fixed: pydantic 1.7 supported
+
+* Fixed: pydantic error not throw